Change how subscriptions handle intraprocess messages
