+++ rclcpp/src/rclcpp/publisher.cpp
@@ -105,7 +105,7 @@ PublisherBase::~PublisherBase()
     // TODO(ivanpauno): should this raise an error?
     RCLCPP_WARN(
       rclcpp::get_logger("rclcpp"),
-      "Intra process manager died before than the publisher.");
+      "Intra process manager died before than a publisher.");
     return;
   }
   ipm->remove_publisher(intra_process_publisher_id_);
