+++ rclcpp/test/test_intra_process_manager.cpp
@@ -86,7 +86,7 @@ public:
     return mapped_ring_buffer::MappedRingBuffer<
       T,
       typename Publisher<T, Alloc>::MessageAlloc
-      >::make_shared(size, allocator_);
+    >::make_shared(size, allocator_);
   }
 
   std::shared_ptr<MessageAlloc> get_allocator()
@@ -129,7 +129,7 @@ public:
 
 // Prevent rclcpp/publisher_base.hpp and rclcpp/subscription.hpp from being imported.
 #define RCLCPP__PUBLISHER_BASE_HPP_
-#define RCLCPP__SUBSCRIPTION_HPP_
+#define RCLCPP__SUBSCRIPTION_BASE_HPP_
 #define RCLCPP_BUILDING_LIBRARY 1
 // Force ipm to use our mock publisher class.
 #define Publisher mock::Publisher
